Panel with current projects
Implement a new panel on the overview page containing the currently active
projects.  The user can quickly change the status of these projects.

Refactor the handling of stylesheets and scripts on overview page: now each
panel may define its own stylesheet and script which is loaded along the
overview page.

Also change the icon for list view and update the changelog text.