Helpdesks for end users
Implement helpdesks which are not associated to an organization and, thus,
suitable for end users.  These helpdesks do not show an overview of all tickets
in frontend.  The user may only create new tickets and see its created ticket
in show view.

Also fix some minor issues such as wrong height of toolbars or boilerplates in
textareas in frontend.