Staff implementation / Data conversion issues
Fix a bug in Grails that empty strings are not converted correctly to
BigDecimal.  In the default Grails converter a NPE occurs.

Add converters for LocalDate, LocalDateTime and LocalTime.

Remove casting from params to BindingResult because it didn't update
some values.

Remove property phoneOther from search fields of class Staff.