Handle empty lists in find actions
Handle the case of an empty list in find actions which causes respond()
to set the expected model variable (e. g. personList) to null instead
of an empty list.  This is because respond() cannot obtain the type of
items of an empty list and, thus, stores it the empty list in the model
under name emptyCollection.

Also add corresponding unit tests.