Fix client scripting / person picture upload
Upgrade bootstrap-datepicker and bootstrap-fileinput dependencies to
their latest versions.  Change the code used to display the fileinput
widget.  Upgrade jQuery to the latest version.

Use module system from another project.  Change all the client scripting
code to register and require modules.

Implement a task to rewrite bootstrap-fileinput language files because
they need special handling.  Also, add a conversion from LESS to SCSS
for bootstrap-datepicker stylesheets because they are available in LESS,
only.

Move all code for fileinput widgets to a separate SCSS file and include
it only where needed.

Change the method for PersonController.update to POST because multipart
requests are not handled when using PUT method.  Add an associated unit
test.

Handle a bug that addresses of organization were not loaded via AJAX.

Fix an issue that the full number of invoicing transaction was not
displayed in the title.